requirements.txt
# Core Python Packages
numpy>=1.24.0
pillow>=10.0.0
asyncio-extensions>=0.1.0

# 3D Graphics and Rendering
PyOpenGL>=3.1.7
moderngl>=5.8.0
pyrr>=0.10.3  # 3D math utilities
glfw>=2.6.0   # Window management

# VRM and 3D Model Support
pygltflib>=1.16.0
trimesh>=4.0.0
scipy>=1.11.0

# GUI and Window Management
tkinter  # Built into Python
pywin32>=306; sys_platform=="win32"
pillow>=10.0.0

# AI and Language Models
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.16.0
google-cloud-speech>=2.21.0
ollama>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Voice and Audio
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
azure-cognitiveservices-speech>=1.34.0
openai-whisper>=20231117
sounddevice>=0.4.6
librosa>=0.10.0

# Real-time Communication  
livekit>=0.10.0
livekit-api>=0.5.0
websockets>=12.0
aiohttp>=3.9.0

# System Integration
psutil>=5.9.0
pyautogui>=0.9.54
keyboard>=0.13.5
mouse>=0.7.1

# Configuration and Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
rich>=13.7.0
typer>=0.9.0
watchdog>=3.0.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
